test: split layout store coverage by concern
