+++ Tests/OpenUsageTests/LayoutStoreTests.swift
@@ -25,580 +25,7 @@ final class LayoutStoreTests: XCTestCase {
         XCTAssertTrue(store.placed.isEmpty)
     }
 
-    // MARK: - Undo (#603)
-
-    func testUndoRestoresRemovedMetricToSamePosition() {
-        let store = makeStore("UndoRestoresPosition")
-        // Enable Claude's full set so the order is well-defined and the removed metric has neighbours.
-        for id in ["claude.session", "claude.weekly", "claude.extra", "claude.today"] {
-            store.setMetricEnabled(id, true)
-        }
-        let orderBefore = store.orderedSupportedMetrics(for: "claude").map(\.id)
-        let enabledBefore = store.placed.filter { $0.descriptorID.hasPrefix("claude.") }.map(\.descriptorID)
-
-        // Remove a middle metric, then undo it.
-        store.setMetricEnabled("claude.weekly", false)
-        XCTAssertFalse(store.isMetricEnabled("claude.weekly"))
-        XCTAssertTrue(store.canUndo)
-
-        XCTAssertTrue(store.undo())
-
-        // Re-enabled and back in its exact slot, with the enabled placed order unchanged.
-        XCTAssertTrue(store.isMetricEnabled("claude.weekly"))
-        XCTAssertEqual(store.orderedSupportedMetrics(for: "claude").map(\.id), orderBefore)
-        XCTAssertEqual(
-            store.placed.filter { $0.descriptorID.hasPrefix("claude.") }.map(\.descriptorID),
-            enabledBefore
-        )
-    }
-
-    func testUndoReversesEnable() {
-        let store = makeStore("UndoEnable")
-        // cursor.credits is not in DefaultLayout.metricIDs, so it starts disabled in the mock.
-        XCTAssertFalse(store.isMetricEnabled("cursor.credits"))
-
-        store.setMetricEnabled("cursor.credits", true)
-        XCTAssertTrue(store.isMetricEnabled("cursor.credits"))
-        XCTAssertTrue(store.canUndo)
-
-        XCTAssertTrue(store.undo())
-        XCTAssertFalse(store.isMetricEnabled("cursor.credits"), "undo turns an enabled metric back off")
-    }
-
-    func testUndoReversesMetricReorder() {
-        let store = makeStore("UndoReorderMetric")
-        for id in ["claude.session", "claude.weekly", "claude.extra", "claude.today"] {
-            store.setMetricEnabled(id, true)
-        }
-        let orderBefore = store.orderedSupportedMetrics(for: "claude").map(\.id)
-
-        XCTAssertTrue(store.reorderMetric(dragged: "claude.today", target: "claude.session", in: "claude"))
-        XCTAssertNotEqual(store.orderedSupportedMetrics(for: "claude").map(\.id), orderBefore)
-
-        XCTAssertTrue(store.undo())
-        XCTAssertEqual(store.orderedSupportedMetrics(for: "claude").map(\.id), orderBefore,
-                       "undo restores the exact prior metric order")
-    }
-
-    func testUndoReversesProviderReorder() {
-        let store = makeStore("UndoReorderProvider")
-        let orderBefore = store.customizeGroups.map(\.provider.id)
-
-        XCTAssertTrue(store.reorderProvider(dragged: "cursor", target: "claude"))
-        XCTAssertNotEqual(store.customizeGroups.map(\.provider.id), orderBefore)
-
-        XCTAssertTrue(store.undo())
-        XCTAssertEqual(store.customizeGroups.map(\.provider.id), orderBefore,
-                       "undo restores the exact prior provider order")
-    }
-
-    func testUndoReversesPinAndUnpin() {
-        let store = makeStore("UndoPin")
-        // cursor.usage is enabled by default but not pinned (cursor's default pins aren't in the mock).
-        XCTAssertTrue(store.isMetricEnabled("cursor.usage"))
-        XCTAssertFalse(store.isPinned("cursor.usage"))
-
-        // Pin, then undo → back to unpinned.
-        store.setPinned(true, for: "cursor.usage")
-        XCTAssertTrue(store.isPinned("cursor.usage"))
-        XCTAssertTrue(store.undo())
-        XCTAssertFalse(store.isPinned("cursor.usage"), "undo reverses a pin")
-
-        // Unpin a default-pinned metric, then undo → back to pinned.
-        XCTAssertTrue(store.isPinned("claude.session"))
-        store.setPinned(false, for: "claude.session")
-        XCTAssertFalse(store.isPinned("claude.session"))
-        XCTAssertTrue(store.undo())
-        XCTAssertTrue(store.isPinned("claude.session"), "undo reverses an unpin")
-    }
-
-    func testUndoReversesExpandedMove() {
-        let store = makeStore("UndoExpandedMove")
-        // claude.session stays above the fold by default (not in DefaultLayout.expandedMetricIDs).
-        XCTAssertFalse(store.expandedMetricIDs.contains("claude.session"))
-
-        XCTAssertTrue(moveMetric("claude.session", expanded: true, in: store))
-        XCTAssertTrue(store.expandedMetricIDs.contains("claude.session"))
-
-        XCTAssertTrue(store.undo())
-        XCTAssertFalse(store.expandedMetricIDs.contains("claude.session"), "undo moves the metric back above the caret")
-    }
-
-    func testUndoDoesNotRestoreProviderCardCaretState() {
-        // Provider card expand/collapse is transient view state, not a layout edit, so undo must leave
-        // the caret where the user last put it — not rewind it to whatever was open when the undone
-        // step was recorded. Regression test for the snapshot wrongly capturing expandedProviderIDs.
-        let store = makeStore("UndoLeavesProviderCaret")
-        XCTAssertFalse(store.isProviderExpanded("codex"))
-
-        // Open Codex's card, then make an undoable layout edit. The pre-edit snapshot must not capture
-        // the open caret.
-        XCTAssertTrue(store.setProviderExpanded(true, for: "codex"))
-        XCTAssertTrue(store.isProviderExpanded("codex"))
-        store.setMetricEnabled("cursor.credits", true)
-        XCTAssertTrue(store.canUndo)
-
-        // Collapse the card after the step was recorded, then undo the enable.
-        XCTAssertTrue(store.setProviderExpanded(false, for: "codex"))
-        XCTAssertFalse(store.isProviderExpanded("codex"))
-
-        XCTAssertTrue(store.undo())
-        XCTAssertFalse(store.isMetricEnabled("cursor.credits"))
-        XCTAssertFalse(store.isProviderExpanded("codex"), "undo must not restore provider card caret state")
-    }
-
-    func testUndoWalksBackMultipleMixedSteps() {
-        let store = makeStore("UndoMultiStep")
-        // Distinct, real changes: enable an off metric, pin an unpinned one, remove an on metric.
-        store.setMetricEnabled("cursor.credits", true)  // step 1: enable
-        store.setPinned(true, for: "cursor.usage")      // step 2: pin
-        store.setMetricEnabled("claude.session", false) // step 3: remove
-
-        // Walk back in reverse order, one step per ⌘Z.
-        XCTAssertTrue(store.undo())                      // undo remove
-        XCTAssertTrue(store.isMetricEnabled("claude.session"))
-        XCTAssertTrue(store.isPinned("cursor.usage"))
-
-        XCTAssertTrue(store.undo())                      // undo pin
-        XCTAssertFalse(store.isPinned("cursor.usage"))
-        XCTAssertTrue(store.isMetricEnabled("cursor.credits"))
-
-        XCTAssertTrue(store.undo())                      // undo enable
-        XCTAssertFalse(store.isMetricEnabled("cursor.credits"))
-
-        XCTAssertFalse(store.canUndo)
-        XCTAssertFalse(store.undo())
-    }
-
-    func testUndoIsNotItselfRecorded() {
-        // Applying an undo must not push a new step — otherwise ⌘Z would ping-pong forever.
-        let store = makeStore("UndoNotRecorded")
-        store.setMetricEnabled("cursor.credits", true)
-        XCTAssertTrue(store.canUndo)
-
-        XCTAssertTrue(store.undo())
-        XCTAssertFalse(store.canUndo, "undo leaves nothing new to undo")
-    }
-
-    func testUndoStackIsCappedAtMaxDepth() {
-        let store = makeStore("UndoMaxDepth")
-        // Drive more distinct, recordable changes than the cap by toggling a pin on and off repeatedly.
-        store.setMetricEnabled("claude.weekly", true)
-        var pinned = false
-        for _ in 0..<(LayoutUndoHistory.maxDepth + 10) {
-            pinned.toggle()
-            store.setPinned(pinned, for: "claude.weekly")
-        }
-        // Undo can only walk back the cap's worth of steps, then stops.
-        var steps = 0
-        while store.undo() { steps += 1 }
-        XCTAssertEqual(steps, LayoutUndoHistory.maxDepth)
-    }
-
-    func testNoOpActionDoesNotRecordUndoStep() {
-        let store = makeStore("UndoNoOp")
-        store.setMetricEnabled("cursor.credits", true)  // one real step
-        // Re-enabling an already-on metric, or a self-target reorder, changes nothing → no step.
-        store.setMetricEnabled("cursor.credits", true)
-        store.reorderMetric(dragged: "claude.weekly", target: "claude.weekly", in: "claude")
-
-        // Exactly one undoable step (the original enable).
-        XCTAssertTrue(store.undo())
-        XCTAssertFalse(store.canUndo)
-    }
-
-    func testUndoWithEmptyHistoryIsNoOp() {
-        let store = makeStore("UndoEmpty")
-        let before = store.placed.map(\.descriptorID)
-
-        XCTAssertFalse(store.canUndo)
-        XCTAssertFalse(store.undo())
-        XCTAssertEqual(store.placed.map(\.descriptorID), before)
-    }
-
-    func testResetToDefaultClearsUndoHistory() {
-        let store = makeStore("UndoResetAllClears")
-        store.setMetricEnabled("claude.weekly", true)
-        store.setMetricEnabled("claude.weekly", false)
-        XCTAssertTrue(store.canUndo)
-
-        store.resetToDefault()
-
-        XCTAssertFalse(store.canUndo)
-        XCTAssertFalse(store.undo())
-    }
-
-    func testResetProviderClearsUndoHistory() {
-        let store = makeStore("UndoResetProviderClears")
-        store.setMetricEnabled("cursor.credits", true)
-        store.setMetricEnabled("cursor.requests", true)
-        XCTAssertTrue(store.canUndo)
-
-        store.resetProvider("claude")
-
-        // Snapshots are whole-layout, so a reset (its own deliberate action) drops the entire stack.
-        XCTAssertFalse(store.canUndo)
-        XCTAssertFalse(store.undo())
-    }
-
-    func testDirectRemoveDoesNotRecordUndo() {
-        // The low-level `remove(_:)` (used by drag teardown and tests) is not a user-facing seam, so it
-        // doesn't feed the undo stack — only the wrapped mutations (setMetricEnabled, reorder, pin) do.
-        let store = LayoutStore(
-            registry: .mock,
-            defaults: makeDefaults("UndoDirectRemove"),
-            storageKey: "layout",
-            defaultMetricIDs: ["claude.weekly"]
-        )
-        guard let widget = store.placed.first(where: { $0.descriptorID == "claude.weekly" }) else {
-            return XCTFail("metric was not placed")
-        }
-
-        store.remove(widget.id)
-
-        XCTAssertFalse(store.canUndo)
-    }
-
-    func testSavedEmptyLayoutDoesNotRestoreDefaults() {
-        let defaults = makeDefaults("EmptyLayout")
-        let store = LayoutStore(registry: .mock, defaults: defaults, storageKey: "layout")
-
-        for widget in store.placed {
-            store.remove(widget.id)
-        }
-
-        let reloaded = LayoutStore(registry: .mock, defaults: defaults, storageKey: "layout")
-        XCTAssertTrue(reloaded.placed.isEmpty)
-    }
-
-    func testUnreadableStoredLayoutIsNotMistakenForFreshInstall() throws {
-        let defaults = makeDefaults("UnreadableExistingLayout")
-        defaults.set(Data("not valid layout data".utf8), forKey: "layout")
-
-        let store = LayoutStore(
-            registry: .mock,
-            defaults: defaults,
-            storageKey: "layout",
-            defaultExpandedMetricIDs: ["claude.weekly"]
-        )
-
-        XCTAssertFalse(
-            store.expandedMetricIDs.contains("claude.weekly"),
-            "present but damaged data is still an existing layout, so fresh-only defaults must stay off"
-        )
-        XCTAssertNoThrow(
-            try JSONDecoder().decode(
-                [PlacedWidget].self,
-                from: XCTUnwrap(defaults.data(forKey: "layout"))
-            ),
-            "the recovered default layout replaces unreadable storage instead of failing every launch"
-        )
-    }
-
-    func testUnreadableSeedMarkerKeepsExistingUserBaseline() {
-        let defaults = makeDefaults("UnreadableSeedMarker")
-        saveStored([PlacedWidget(descriptorID: "claude.session")], forKey: "layout", in: defaults)
-        defaults.set(Data("not valid seeded-default data".utf8), forKey: "layout.seededDefaults")
-
-        let store = LayoutStore(
-            registry: .mock,
-            defaults: defaults,
-            storageKey: "layout",
-            defaultMetricIDs: ["claude.session", "claude.weekly"],
-            migrationBaselineMetricIDs: ["claude.session", "claude.weekly"]
-        )
-
-        XCTAssertEqual(store.placed.map(\.descriptorID), ["claude.session"])
-    }
-
-    func testExistingLayoutAutoSeedsOnlyDefaultsAddedAfterBaseline() {
-        let defaults = makeDefaults("SeedNewDefault")
-        saveStored([PlacedWidget(descriptorID: "claude.session")], forKey: "layout", in: defaults)
-
-        let store = LayoutStore(
-            registry: .mock,
-            defaults: defaults,
-            storageKey: "layout",
-            defaultMetricIDs: ["claude.session", "claude.weekly", "claude.today"],
-            migrationBaselineMetricIDs: ["claude.session", "claude.weekly"]
-        )
-
-        XCTAssertEqual(store.placed.map(\.descriptorID), ["claude.session", "claude.today"])
-        XCTAssertFalse(store.isMetricEnabled("claude.weekly"), "baseline defaults the user already removed stay off")
-    }
-
-    func testDisablingAutoSeededDefaultDoesNotReAddOnReload() {
-        let defaults = makeDefaults("SeedOnce")
-        saveStored([PlacedWidget(descriptorID: "claude.session")], forKey: "layout", in: defaults)
-        let store = LayoutStore(
-            registry: .mock,
-            defaults: defaults,
-            storageKey: "layout",
-            defaultMetricIDs: ["claude.session", "claude.today"],
-            migrationBaselineMetricIDs: ["claude.session"]
-        )
-        guard let seeded = store.placed.first(where: { $0.descriptorID == "claude.today" }) else {
-            return XCTFail("new default was not seeded")
-        }
-
-        store.remove(seeded.id)
-
-        let reloaded = LayoutStore(
-            registry: .mock,
-            defaults: defaults,
-            storageKey: "layout",
-            defaultMetricIDs: ["claude.session", "claude.today"],
-            migrationBaselineMetricIDs: ["claude.session"]
-        )
-        XCTAssertEqual(reloaded.placed.map(\.descriptorID), ["claude.session"])
-    }
-
-    func testFreshLayoutTreatsCurrentDefaultsAsAlreadySeeded() {
-        let defaults = makeDefaults("FreshSeeded")
-        let store = LayoutStore(
-            registry: .mock,
-            defaults: defaults,
-            storageKey: "layout",
-            defaultMetricIDs: ["claude.session", "claude.today"],
-            migrationBaselineMetricIDs: []
-        )
-        guard let today = store.placed.first(where: { $0.descriptorID == "claude.today" }) else {
-            return XCTFail("fresh store did not include all current defaults")
-        }
-
-        store.remove(today.id)
-
-        let reloaded = LayoutStore(
-            registry: .mock,
-            defaults: defaults,
-            storageKey: "layout",
-            defaultMetricIDs: ["claude.session", "claude.today"],
-            migrationBaselineMetricIDs: []
-        )
-        XCTAssertEqual(reloaded.placed.map(\.descriptorID), ["claude.session"])
-    }
-
-    func testAutoSeedingIgnoresUnknownDefaultIDs() {
-        let defaults = makeDefaults("UnknownSeed")
-        saveStored([PlacedWidget](), forKey: "layout", in: defaults)
-
-        let store = LayoutStore(
-            registry: .mock,
-            defaults: defaults,
-            storageKey: "layout",
-            defaultMetricIDs: ["missing.metric", "claude.session"],
-            migrationBaselineMetricIDs: []
-        )
-
-        XCTAssertEqual(store.placed.map(\.descriptorID), ["claude.session"])
-    }
-
-    func testExistingLayoutEnablesDefaultExpandedOptionalBelowCaret() {
-        let defaults = makeDefaults("LegacyEnableExpanded")
-        saveStored([PlacedWidget(descriptorID: "cursor.usage")], forKey: "layout", in: defaults)
-        let store = LayoutStore(
-            registry: .mock,
-            defaults: defaults,
-            storageKey: "layout",
-            defaultMetricIDs: ["cursor.usage"],
-            migrationBaselineMetricIDs: ["cursor.usage"],
-            defaultExpandedMetricIDs: ["cursor.requests"]
-        )
-
-        XCTAssertFalse(store.isMetricEnabled("cursor.requests"))
-        XCTAssertFalse(store.expandedMetricIDs.contains("cursor.requests"))
-
-        store.setMetricEnabled("cursor.requests", true)
-
-        XCTAssertTrue(store.isMetricEnabled("cursor.requests"))
-        XCTAssertTrue(store.expandedMetricIDs.contains("cursor.requests"))
-    }
-
-    func testNewlySeededDefaultExpandedMetricEntersBelowCaretForExistingLayout() {
-        let defaults = makeDefaults("SeedNewExpanded")
-        // An existing layout from before the new metric shipped, with a saved expanded set that can't
-        // know about it yet.
-        saveStored([PlacedWidget(descriptorID: "claude.session")], forKey: "layout", in: defaults)
-        defaults.set(["claude.weekly"], forKey: "layout.expandedMetrics")
-
-        let store = LayoutStore(
-            registry: .mock,
-            defaults: defaults,
-            storageKey: "layout",
-            defaultMetricIDs: ["claude.session", "claude.today"],
-            migrationBaselineMetricIDs: ["claude.session"],
-            defaultExpandedMetricIDs: ["claude.today"]
-        )
-
-        // The new default is auto-enabled by migration AND tucked below the caret, not surfaced primary.
-        XCTAssertTrue(store.isMetricEnabled("claude.today"))
-        XCTAssertTrue(store.expandedMetricIDs.contains("claude.today"))
-        // A metric the user already lived with stays always-shown.
-        XCTAssertFalse(store.expandedMetricIDs.contains("claude.session"))
-
-        // The new expanded membership persists across reloads.
-        let reloaded = LayoutStore(
-            registry: .mock,
-            defaults: defaults,
-            storageKey: "layout",
-            defaultMetricIDs: ["claude.session", "claude.today"],
-            migrationBaselineMetricIDs: ["claude.session"],
-            defaultExpandedMetricIDs: ["claude.today"]
-        )
-        XCTAssertTrue(reloaded.expandedMetricIDs.contains("claude.today"))
-    }
-
-    func testMigrationPersistKeepsLegacyOptionalMetricExpandOnEnableAfterReload() {
-        let defaults = makeDefaults("SeedExpandedKeepsFallback")
-        // Legacy layout: predates the expanded feature (no saved expanded set).
-        saveStored([PlacedWidget(descriptorID: "cursor.usage")], forKey: "layout", in: defaults)
-
-        let args: (UserDefaults) -> LayoutStore = { d in
-            LayoutStore(
-                registry: .mock,
-                defaults: d,
-                storageKey: "layout",
-                defaultMetricIDs: ["cursor.usage", "claude.today"],
-                migrationBaselineMetricIDs: ["cursor.usage"],
-                // claude.today is a brand-new default (auto-enabled + tucked, persisting an expanded set);
-                // cursor.requests is an optional default-expanded metric the user hasn't enabled yet.
-                defaultExpandedMetricIDs: ["claude.today", "cursor.requests"]
-            )
-        }
-
-        // First launch performs the migration and persists the expanded set.
-        _ = args(defaults)
-
-        // Second launch now sees a saved expanded set — the legacy optional metric must still enter below
-        // the caret when first enabled (regression: persisting the migration zeroed the on-enable queue).
-        let reloaded = args(defaults)
-        XCTAssertFalse(reloaded.expandedMetricIDs.contains("cursor.requests"))
-        reloaded.setMetricEnabled("cursor.requests", true)
-        XCTAssertTrue(reloaded.expandedMetricIDs.contains("cursor.requests"))
-    }
-
-    func testConsumedExpandOnEnableStaysConsumedAcrossRelaunch() {
-        let defaults = makeDefaults("ExpandOnEnablePersists")
-        saveStored([PlacedWidget(descriptorID: "cursor.usage")], forKey: "layout", in: defaults)
-
-        let args: (UserDefaults) -> LayoutStore = { d in
-            LayoutStore(
-                registry: .mock,
-                defaults: d,
-                storageKey: "layout",
-                defaultMetricIDs: ["cursor.usage"],
-                migrationBaselineMetricIDs: ["cursor.usage"],
-                defaultExpandedMetricIDs: ["cursor.requests"]
-            )
-        }
-
-        // The user drags the still-disabled optional metric above the divider — an explicit placement
-        // that consumes its expand-on-enable default.
-        let store = args(defaults)
-        XCTAssertTrue(store.reorderMetric(dragged: "cursor.requests", target: "cursor.usage", in: "cursor"))
-        XCTAssertFalse(store.expandedMetricIDs.contains("cursor.requests"))
-
-        // After a relaunch the consumed default must stay consumed — enabling it leaves it above the fold
-        // (regression: the queue was recomputed each launch and resurrected the consumed entry).
-        let reloaded = args(defaults)
-        reloaded.setMetricEnabled("cursor.requests", true)
-        XCTAssertTrue(reloaded.isMetricEnabled("cursor.requests"))
-        XCTAssertFalse(reloaded.expandedMetricIDs.contains("cursor.requests"))
-    }
-
-    func testExplicitDividerMoveOverridesDefaultExpandedOnEnable() {
-        let defaults = makeDefaults("LegacyEnableExpandedOverride")
-        saveStored([PlacedWidget(descriptorID: "cursor.usage")], forKey: "layout", in: defaults)
-        let store = LayoutStore(
-            registry: .mock,
-            defaults: defaults,
-            storageKey: "layout",
-            defaultMetricIDs: ["cursor.usage"],
-            migrationBaselineMetricIDs: ["cursor.usage"],
-            defaultExpandedMetricIDs: ["cursor.requests"]
-        )
-        let divider = "cursor::expanded-divider"
-
-        XCTAssertTrue(store.applyMetricDividerOrder([
-            "cursor.usage",
-            "cursor.requests",
-            divider,
-            "cursor.credits",
-            "cursor.today"
-        ], dragged: "cursor.requests", dividerID: divider, in: "cursor"))
-        store.setMetricEnabled("cursor.requests", true)
-
-        XCTAssertTrue(store.isMetricEnabled("cursor.requests"))
-        XCTAssertFalse(store.expandedMetricIDs.contains("cursor.requests"))
-
-        let reloaded = LayoutStore(
-            registry: .mock,
-            defaults: defaults,
-            storageKey: "layout",
-            defaultMetricIDs: ["cursor.usage"],
-            migrationBaselineMetricIDs: ["cursor.usage"],
-            defaultExpandedMetricIDs: ["cursor.requests"]
-        )
-        reloaded.setMetricEnabled("cursor.requests", true)
-        XCTAssertFalse(reloaded.expandedMetricIDs.contains("cursor.requests"))
-    }
-
-    func testPrimaryDividerReorderDoesNotConsumeHiddenDefaultExpandedOnEnable() {
-        let defaults = makeDefaults("LegacyPrimaryReorderKeepsFallback")
-        saveStored([
-            PlacedWidget(descriptorID: "cursor.usage"),
-            PlacedWidget(descriptorID: "cursor.today")
-        ], forKey: "layout", in: defaults)
-        let store = LayoutStore(
-            registry: .mock,
-            defaults: defaults,
-            storageKey: "layout",
-            defaultMetricIDs: ["cursor.usage", "cursor.today"],
-            migrationBaselineMetricIDs: ["cursor.usage", "cursor.today"],
-            defaultExpandedMetricIDs: ["cursor.requests"]
-        )
-        let divider = "cursor::expanded-divider"
-
-        XCTAssertTrue(store.applyMetricDividerOrder([
-            "cursor.today",
-            "cursor.usage",
-            divider
-        ], dragged: "cursor.today", dividerID: divider, in: "cursor"))
-        store.setMetricEnabled("cursor.requests", true)
-
-        XCTAssertTrue(store.expandedMetricIDs.contains("cursor.requests"))
-    }
-
-    func testCustomizeReorderDoesNotConsumeUnmovedDisabledExpandOnEnable() {
-        let defaults = makeDefaults("CustomizePrimaryReorderKeepsUnmovedFallback")
-        saveStored([
-            PlacedWidget(descriptorID: "cursor.usage"),
-            PlacedWidget(descriptorID: "cursor.today")
-        ], forKey: "layout", in: defaults)
-        let store = LayoutStore(
-            registry: .mock,
-            defaults: defaults,
-            storageKey: "layout",
-            defaultMetricIDs: ["cursor.usage", "cursor.today"],
-            migrationBaselineMetricIDs: ["cursor.usage", "cursor.today"],
-            defaultExpandedMetricIDs: ["cursor.requests"]
-        )
-        let divider = "cursor::expanded-divider"
-
-        // Customize passes the full metric list (metricOrderWithDivider includes the disabled
-        // cursor.requests before the divider) even when only reordering primary rows. The dragged
-        // metric is cursor.today, not cursor.requests — so cursor.requests' below-caret default must
-        // survive the reorder and still place it below the caret when later enabled.
-        XCTAssertTrue(store.applyMetricDividerOrder([
-            "cursor.today",
-            "cursor.usage",
-            "cursor.requests",
-            divider
-        ], dragged: "cursor.today", dividerID: divider, in: "cursor"))
-        store.setMetricEnabled("cursor.requests", true)
-
-        XCTAssertTrue(store.expandedMetricIDs.contains("cursor.requests"))
-    }
+    // MARK: - Placement, ordering, and defaults
 
     func testAddAndResetCancelDragState() {
         let store = makeStore("CancelDrag")
@@ -830,429 +257,6 @@ final class LayoutStoreTests: XCTestCase {
         XCTAssertEqual(reloaded.placed.map(\.descriptorID), ["claude.session"])
     }
 
-    // MARK: - Expanded ("Shown on expand") membership
-
-    func testDividerDragMovesMetricBelowDividerAndPersists() {
-        let defaults = makeDefaults("ExpandMove")
-        // Hermetic: start with nothing below the caret (independent of DefaultLayout's seeding).
-        let store = LayoutStore(registry: .mock, defaults: defaults, storageKey: "layout", defaultExpandedMetricIDs: [])
-        guard let first = store.orderedSupportedMetrics(for: "claude").map(\.id).first else {
-            return XCTFail("missing Claude metrics")
-        }
-        XCTAssertFalse(store.expandedMetricIDs.contains(first))
-
-        XCTAssertTrue(moveMetric(first, expanded: true, in: store))
-        XCTAssertTrue(store.expandedMetricIDs.contains(first))
-
-        let group = store.customizeGroups.first { $0.provider.id == "claude" }
-        XCTAssertEqual(group?.expandedMetrics.map(\.id).first, first)
-        XCTAssertFalse(group?.alwaysShownMetrics.map(\.id).contains(first) ?? true)
-
-        let reloaded = LayoutStore(registry: .mock, defaults: defaults, storageKey: "layout")
-        XCTAssertTrue(reloaded.expandedMetricIDs.contains(first))
-    }
-
-    func testDividerDragIsNoOpWhenAlreadyInSection() {
-        let store = LayoutStore(registry: .mock, defaults: makeDefaults("ExpandNoOp"), storageKey: "layout", defaultExpandedMetricIDs: [])
-        guard let id = store.orderedSupportedMetrics(for: "claude").map(\.id).first else {
-            return XCTFail("missing Claude metrics")
-        }
-        XCTAssertFalse(moveMetric(id, expanded: false, in: store), "already always-shown")
-        XCTAssertTrue(moveMetric(id, expanded: true, in: store))
-        XCTAssertFalse(moveMetric(id, expanded: true, in: store), "already expanded")
-    }
-
-    func testDraggingMetricOntoExpandedRowTucksItAway() {
-        let store = LayoutStore(
-            registry: .mock,
-            defaults: makeDefaults("DragAcross"),
-            storageKey: "layout",
-            defaultMetricIDs: ["cursor.usage", "cursor.today"],
-            defaultExpandedMetricIDs: []
-        )
-        let ids = store.orderedSupportedMetrics(for: "cursor").map(\.id)
-        guard ids.count >= 2, let dragged = ids.first, let target = ids.last else {
-            return XCTFail("need at least two Cursor metrics")
-        }
-        XCTAssertTrue(moveMetric(target, expanded: true, in: store))
-        XCTAssertFalse(store.expandedMetricIDs.contains(dragged))
-
-        XCTAssertTrue(store.reorderMetric(dragged: dragged, target: target, in: "cursor"))
-
-        XCTAssertTrue(store.expandedMetricIDs.contains(dragged), "dropping onto an expanded row moves the dragged row across")
-        let expanded = store.customizeGroups.first { $0.provider.id == "cursor" }?.expandedMetrics.map(\.id) ?? []
-        XCTAssertTrue(expanded.contains(dragged) && expanded.contains(target))
-    }
-
-    func testDraggingExpandedMetricOntoAlwaysShownRowBringsItBack() {
-        let store = LayoutStore(
-            registry: .mock,
-            defaults: makeDefaults("DragBack"),
-            storageKey: "layout",
-            defaultMetricIDs: ["cursor.usage", "cursor.today"],
-            defaultExpandedMetricIDs: []
-        )
-        let ids = store.orderedSupportedMetrics(for: "cursor").map(\.id)
-        guard ids.count >= 2, let target = ids.first, let dragged = ids.last else {
-            return XCTFail("need at least two Cursor metrics")
-        }
-        XCTAssertTrue(moveMetric(dragged, expanded: true, in: store))
-
-        XCTAssertTrue(store.reorderMetric(dragged: dragged, target: target, in: "cursor"))
-        XCTAssertFalse(store.expandedMetricIDs.contains(dragged), "dropping onto an always-shown row brings the dragged row back")
-    }
-
-    func testApplyingDividerOrderMovesMetricBelowFold() {
-        let store = LayoutStore(
-            registry: .mock,
-            defaults: makeDefaults("DividerDown"),
-            storageKey: "layout",
-            defaultMetricIDs: ["cursor.usage", "cursor.today"],
-            defaultExpandedMetricIDs: []
-        )
-        let divider = "cursor::expanded-divider"
-
-        XCTAssertTrue(store.applyMetricDividerOrder([
-            "cursor.usage",
-            divider,
-            "cursor.requests",
-            "cursor.credits",
-            "cursor.today"
-        ], dragged: "cursor.requests", dividerID: divider, in: "cursor"))
-
-        XCTAssertFalse(store.expandedMetricIDs.contains("cursor.usage"))
-        XCTAssertTrue(store.expandedMetricIDs.contains("cursor.requests"))
-    }
-
-    func testApplyingDividerOrderMovesMetricAboveFold() {
-        let store = LayoutStore(
-            registry: .mock,
-            defaults: makeDefaults("DividerUp"),
-            storageKey: "layout",
-            defaultMetricIDs: ["cursor.usage", "cursor.today"],
-            defaultExpandedMetricIDs: []
-        )
-        let divider = "cursor::expanded-divider"
-        XCTAssertTrue(moveMetric("cursor.requests", expanded: true, in: store))
-
-        XCTAssertTrue(store.applyMetricDividerOrder([
-            "cursor.usage",
-            "cursor.requests",
-            divider,
-            "cursor.credits",
-            "cursor.today"
-        ], dragged: "cursor.requests", dividerID: divider, in: "cursor"))
-
-        XCTAssertFalse(store.expandedMetricIDs.contains("cursor.requests"))
-    }
-
-    func testApplyingVisibleDividerOrderKeepsDisabledMetricsInPlace() {
-        let store = LayoutStore(
-            registry: .mock,
-            defaults: makeDefaults("VisibleDividerKeepsDisabled"),
-            storageKey: "layout",
-            defaultMetricIDs: ["cursor.usage", "cursor.today"],
-            defaultExpandedMetricIDs: ["cursor.requests", "cursor.today"]
-        )
-        let divider = "cursor::expanded-divider"
-
-        XCTAssertFalse(store.isMetricEnabled("cursor.credits"))
-        XCTAssertFalse(store.isMetricEnabled("cursor.requests"))
-        XCTAssertTrue(store.applyMetricDividerOrder([
-            "cursor.usage",
-            "cursor.today",
-            divider
-        ], dragged: "cursor.today", dividerID: divider, in: "cursor"))
-        XCTAssertEqual(store.orderedSupportedMetrics(for: "cursor").map(\.id), [
-            "cursor.usage", "cursor.credits", "cursor.today", "cursor.requests"
-        ])
-        XCTAssertFalse(store.expandedMetricIDs.contains("cursor.today"))
-        XCTAssertTrue(store.expandedMetricIDs.contains("cursor.requests"))
-    }
-
-    func testDisabledMetricKeepsExpandedMembership() {
-        let defaults = makeDefaults("DisabledExpanded")
-        let store = LayoutStore(
-            registry: .mock,
-            defaults: defaults,
-            storageKey: "layout",
-            defaultMetricIDs: ["claude.session"],
-            defaultExpandedMetricIDs: []
-        )
-        XCTAssertFalse(store.isMetricEnabled("claude.extra"))
-
-        XCTAssertTrue(moveMetric("claude.extra", expanded: true, in: store))
-        XCTAssertTrue(store.expandedMetricIDs.contains("claude.extra"))
-        XCTAssertFalse(store.isMetricEnabled("claude.extra"))
-
-        let reloaded = LayoutStore(
-            registry: .mock,
-            defaults: defaults,
-            storageKey: "layout",
-            defaultMetricIDs: ["claude.session"],
-            defaultExpandedMetricIDs: []
-        )
-        XCTAssertTrue(reloaded.expandedMetricIDs.contains("claude.extra"))
-    }
-
-    func testFreshLayoutSeedsDefaultExpanded() {
-        let store = LayoutStore(
-            registry: .mock,
-            defaults: makeDefaults("FreshExpanded"),
-            storageKey: "layout",
-            defaultExpandedMetricIDs: ["claude.weekly"]
-        )
-        XCTAssertTrue(store.expandedMetricIDs.contains("claude.weekly"))
-    }
-
-    func testExistingLayoutDoesNotSeedExpanded() {
-        let defaults = makeDefaults("ExistingNoExpand")
-        saveStored([PlacedWidget(descriptorID: "claude.session")], forKey: "layout", in: defaults)
-        let store = LayoutStore(
-            registry: .mock,
-            defaults: defaults,
-            storageKey: "layout",
-            defaultExpandedMetricIDs: ["claude.weekly"]
-        )
-        XCTAssertFalse(store.expandedMetricIDs.contains("claude.weekly"), "an existing layout keeps every metric always-shown")
-    }
-
-    func testResetRestoresDefaultExpanded() {
-        let store = LayoutStore(
-            registry: .mock,
-            defaults: makeDefaults("ResetExpand"),
-            storageKey: "layout",
-            defaultExpandedMetricIDs: ["claude.weekly"]
-        )
-        XCTAssertTrue(moveMetric("claude.weekly", expanded: false, in: store))
-        XCTAssertFalse(store.expandedMetricIDs.contains("claude.weekly"))
-
-        store.resetToDefault()
-        XCTAssertTrue(store.expandedMetricIDs.contains("claude.weekly"))
-    }
-
-    func testInvalidPersistedExpandedIDsAreDropped() {
-        let defaults = makeDefaults("InvalidExpand")
-        saveStored([PlacedWidget(descriptorID: "claude.session")], forKey: "layout", in: defaults)
-        defaults.set(["claude.session", "missing.metric"], forKey: "layout.expandedMetrics")
-
-        let store = LayoutStore(registry: .mock, defaults: defaults, storageKey: "layout")
-        XCTAssertTrue(store.expandedMetricIDs.contains("claude.session"))
-        XCTAssertFalse(store.expandedMetricIDs.contains("missing.metric"))
-    }
-
-    func testDisplayGroupsPartitionEnabledMetrics() {
-        let store = LayoutStore(
-            registry: .mock,
-            defaults: makeDefaults("DisplayPartition"),
-            storageKey: "layout",
-            defaultMetricIDs: ["claude.session", "claude.weekly"],
-            defaultExpandedMetricIDs: []
-        )
-        XCTAssertTrue(store.isMetricEnabled("claude.session"))
-        XCTAssertTrue(store.isMetricEnabled("claude.weekly"))
-
-        XCTAssertTrue(moveMetric("claude.weekly", expanded: true, in: store))
-
-        let group = store.displayGroups.first { $0.provider.id == "claude" }
-        XCTAssertEqual(group?.alwaysShownWidgets.compactMap { store.descriptor(for: $0)?.id }, ["claude.session"])
-        XCTAssertEqual(group?.expandedWidgets.compactMap { store.descriptor(for: $0)?.id }, ["claude.weekly"])
-        XCTAssertEqual(group?.hasExpandedMetrics, true)
-    }
-
-    func testProviderWithOnlyExpandedMetricsStillShowsRows() {
-        // Only session + weekly enabled, both primary to start, so expanding both makes the whole
-        // provider expanded — independent of DefaultLayout's seeding.
-        let store = LayoutStore(
-            registry: .mock,
-            defaults: makeDefaults("AllExpanded"),
-            storageKey: "layout",
-            defaultMetricIDs: ["claude.session", "claude.weekly"],
-            defaultExpandedMetricIDs: []
-        )
-        XCTAssertTrue(moveMetric("claude.session", expanded: true, in: store))
-        XCTAssertTrue(moveMetric("claude.weekly", expanded: true, in: store))
-
-        let group = store.displayGroups.first { $0.provider.id == "claude" }
-        XCTAssertNotNil(group)
-        XCTAssertFalse(group?.alwaysShownWidgets.isEmpty ?? true, "all-expanded metrics are promoted so the card is never empty")
-        XCTAssertTrue(group?.expandedWidgets.isEmpty ?? false)
-    }
-
-    func testProviderExpandedStatePersistsAcrossReload() {
-        let defaults = makeDefaults("ProviderExpanded")
-        let store = LayoutStore(registry: .mock, defaults: defaults, storageKey: "layout")
-
-        XCTAssertTrue(store.setProviderExpanded(true, for: "codex"))
-        XCTAssertTrue(store.isProviderExpanded("codex"))
-
-        let reloaded = LayoutStore(registry: .mock, defaults: defaults, storageKey: "layout")
-        XCTAssertTrue(reloaded.isProviderExpanded("codex"))
-    }
-
-    func testProviderExpandedStateCanCollapseAndPersists() {
-        let defaults = makeDefaults("ProviderCollapsed")
-        let store = LayoutStore(registry: .mock, defaults: defaults, storageKey: "layout")
-        XCTAssertTrue(store.setProviderExpanded(true, for: "codex"))
-        XCTAssertTrue(store.setProviderExpanded(false, for: "codex"))
-
-        let reloaded = LayoutStore(registry: .mock, defaults: defaults, storageKey: "layout")
-        XCTAssertFalse(reloaded.isProviderExpanded("codex"))
-    }
-
-    func testInvalidPersistedExpandedProviderIDsAreDropped() {
-        let defaults = makeDefaults("InvalidProviderExpanded")
-        defaults.set(["codex", "missing"], forKey: "layout.expandedProviders")
-
-        let store = LayoutStore(registry: .mock, defaults: defaults, storageKey: "layout")
-        XCTAssertTrue(store.isProviderExpanded("codex"))
-        XCTAssertFalse(store.isProviderExpanded("missing"))
-    }
-
-    func testResetClearsProviderExpandedState() {
-        let store = LayoutStore(registry: .mock, defaults: makeDefaults("ResetProviderExpanded"), storageKey: "layout")
-        XCTAssertTrue(store.setProviderExpanded(true, for: "codex"))
-
-        store.resetToDefault()
-
-        XCTAssertFalse(store.isProviderExpanded("codex"))
-    }
-
-    func testResetProviderRestoresOneProviderAndLeavesOthersAndOrderUntouched() {
-        let defaults = makeDefaults("ResetOneProvider")
-        let store = LayoutStore(
-            registry: .mock,
-            defaults: defaults,
-            storageKey: "layout",
-            defaultMetricIDs: ["claude.session", "codex.session"],
-            migrationBaselineMetricIDs: [],
-            defaultPinnedMetricIDs: ["claude.session", "codex.session"],
-            defaultExpandedMetricIDs: []
-        )
-
-        // Reorder providers first, so we can prove a per-provider reset leaves provider order alone.
-        store.reorderProvider(dragged: "cursor", target: "claude")
-        let orderBefore = store.customizeGroups.map(\.provider.id)
-
-        // Diverge Claude from its defaults in every dimension a reset should restore.
-        store.setMetricEnabled("claude.weekly", true)
-        store.setPinned(true, for: "claude.weekly")
-        store.setProviderExpanded(true, for: "claude")
-        store.reorderMetric(dragged: "claude.extra", target: "claude.session", in: "claude")
-
-        // Diverge Codex too — a Claude reset must not touch it.
-        store.setMetricEnabled("codex.weekly", true)
-        store.setPinned(true, for: "codex.weekly")
-
-        store.resetProvider("claude")
-
-        // Claude restored: enabled set, metric order, pins, and expanded state back to default.
-        XCTAssertTrue(store.isMetricEnabled("claude.session"))
-        XCTAssertFalse(store.isMetricEnabled("claude.weekly"))
-        XCTAssertTrue(store.isPinned("claude.session"))
-        XCTAssertFalse(store.isPinned("claude.weekly"))
-        XCTAssertFalse(store.isProviderExpanded("claude"))
-        XCTAssertEqual(
-            store.orderedSupportedMetrics(for: "claude").map(\.id),
-            MockData.descriptors(for: "claude").map(\.id)
-        )
-
-        // Codex untouched by a Claude-only reset.
-        XCTAssertTrue(store.isMetricEnabled("codex.weekly"))
-        XCTAssertTrue(store.isPinned("codex.weekly"))
-
-        // Provider order untouched — contents-only reset.
-        XCTAssertEqual(store.customizeGroups.map(\.provider.id), orderBefore)
-    }
-
-    func testResetProviderIsNoOpForUnknownProvider() {
-        let store = makeStore("ResetUnknownProvider")
-        let before = store.placed.map(\.descriptorID)
-        store.resetProvider("nope")
-        XCTAssertEqual(store.placed.map(\.descriptorID), before)
-    }
-
-    // MARK: - Customize master/detail (L1 list + L2 detail)
-
-    func testCustomizeProviderRowsIncludesAllProvidersRegardlessOfEnablement() {
-        // Disable Codex; L1 must still list it (greyed), in the registry's provider order.
-        let store = LayoutStore(
-            registry: .mock,
-            defaults: makeDefaults("RowsIncludeDisabled"),
-            storageKey: "layout",
-            isProviderEnabled: { id in id != "codex" }
-        )
-        XCTAssertEqual(store.customizeProviderRows.map(\.id), MockData.providers.map(\.id))
-        let codex = store.customizeProviderRows.first { $0.id == "codex" }
-        XCTAssertNotNil(codex, "disabled provider stays visible in L1")
-        XCTAssertFalse(codex?.isEnabled ?? true, "disabled provider row reports isEnabled false")
-        XCTAssertTrue(store.customizeProviderRows.first { $0.id == "claude" }?.isEnabled ?? false)
-    }
-
-    func testCustomizeProviderRowsCarriesMetricAndPinnedCounts() {
-        let store = makeStore("RowCounts")
-        for row in store.customizeProviderRows {
-            XCTAssertEqual(row.metricCount, MockData.descriptors(for: row.id).count)
-            XCTAssertEqual(row.pinnedCount, store.pinnedCount(forProvider: row.id))
-        }
-    }
-
-    func testCustomizeDetailReturnsMetricsEvenWhenDisabled() {
-        let store = LayoutStore(
-            registry: .mock,
-            defaults: makeDefaults("DetailWhenDisabled"),
-            storageKey: "layout",
-            isProviderEnabled: { id in id != "codex" }
-        )
-        // customizeGroups drops the disabled provider; customizeDetail does not.
-        XCTAssertNil(store.customizeGroups.first { $0.provider.id == "codex" })
-        let detail = store.customizeDetail(for: "codex")
-        XCTAssertNotNil(detail, "disabled provider still has a detail to render dimmed")
-        XCTAssertEqual(detail?.metrics.map(\.id), store.orderedSupportedMetrics(for: "codex").map(\.id))
-    }
-
-    func testCustomizeDetailSplitsAcrossDivider() {
-        let store = LayoutStore(
-            registry: .mock,
-            defaults: makeDefaults("DetailSplit"),
-            storageKey: "layout",
-            defaultMetricIDs: ["claude.session", "claude.weekly"],
-            defaultExpandedMetricIDs: ["claude.weekly"]
-        )
-        let detail = store.customizeDetail(for: "claude")
-        XCTAssertEqual(detail?.expandedMetrics.map(\.id), ["claude.weekly"])
-        XCTAssertEqual(detail?.alwaysShownMetrics.map(\.id), ["claude.session", "claude.extra", "claude.today"])
-    }
-
-    func testCustomizeDetailIsNilForUnknownProvider() {
-        let store = makeStore("DetailUnknown")
-        XCTAssertNil(store.customizeDetail(for: "nope"))
-    }
-
-    func testMetricCountMatchesRegistryDescriptors() {
-        let store = makeStore("MetricCount")
-        for id in MockData.providers.map(\.id) {
-            XCTAssertEqual(store.metricCount(for: id), MockData.descriptors(for: id).count)
-        }
-        XCTAssertEqual(store.metricCount(for: "missing"), 0)
-    }
-
-    func testCustomizeProviderIDClearsWhenLeavingCustomize() {
-        let store = makeStore("RouteClears")
-        store.screen = .customize
-        store.customizeProviderID = "claude"
-        XCTAssertEqual(store.customizeProviderID, "claude")
-
-        store.screen = .dashboard
-        XCTAssertNil(store.customizeProviderID, "leaving Customize resets the L2 selection back to the list")
-
-        // A direct jump to Settings also clears it — never strand a detail selection on another screen.
-        store.screen = .customize
-        store.customizeProviderID = "codex"
-        store.screen = .settings
-        XCTAssertNil(store.customizeProviderID)
-    }
-
     // MARK: - Share confirmation
 
     /// `clearShareConfirmation` hides the pill immediately and cancels the auto-clear task, so a
@@ -1270,7 +274,7 @@ final class LayoutStoreTests: XCTestCase {
 
     /// Move a metric through the same divider-reorder route used by both dashboard and Customize
     /// drags. Tests use this only when they need to perform that real user action as setup.
-    private func moveMetric(_ descriptorID: String, expanded: Bool, in store: LayoutStore) -> Bool {
+    func moveMetric(_ descriptorID: String, expanded: Bool, in store: LayoutStore) -> Bool {
         guard store.expandedMetricIDs.contains(descriptorID) != expanded,
               let providerID = descriptorID.split(separator: ".", maxSplits: 1).first.map(String.init)
         else { return false }
@@ -1287,18 +291,18 @@ final class LayoutStoreTests: XCTestCase {
         )
     }
 
-    private func makeStore(_ name: String) -> LayoutStore {
+    func makeStore(_ name: String) -> LayoutStore {
         LayoutStore(registry: .mock, defaults: makeDefaults(name), storageKey: "layout")
     }
 
-    private func makeDefaults(_ name: String) -> UserDefaults {
+    func makeDefaults(_ name: String) -> UserDefaults {
         let suiteName = "OpenUsageTests.LayoutStore.\(name).\(UUID().uuidString)"
         let defaults = UserDefaults(suiteName: suiteName)!
         defaults.removePersistentDomain(forName: suiteName)
         return defaults
     }
 
-    private func saveStored<T: Encodable>(_ value: T, forKey key: String, in defaults: UserDefaults) {
+    func saveStored<T: Encodable>(_ value: T, forKey key: String, in defaults: UserDefaults) {
         defaults.set(try! JSONEncoder().encode(value), forKey: key)
     }
 }
